Project is executable from start to finish and outputs the atlas of the results. It got stable around 120 epochs, with final results being:
Training Step: 7200  | total loss: 0.08882 | time: 4.374s
| AdaDelta | epoch: 180 | loss: 0.08882 - R2: -0.0091 | val_loss: 0.09503 - val_R2: -0.0517 -- iter: 1259/1259

Also uploaded an image which clearly reflects the R2 ~0.0 obtained (The "same" for all inputs).

Added a few more Fully connected layers, a pair in the middle of the autoencoder and one between the decoder's last convolution and the final output layer, to see how it behaves.